microservices/vehicle-yolo/requirements.txt
flask
pillow
numpy
ultralytics==8.3.230
torch==2.2.0+cpu
torchvision==0.17.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu